+++ js/main.js
@@ -1,6 +1,6 @@
 
-import { photoValues } from './until';
-import { createImage } from './data';
+import { photoValues } from './data';
+import { createImage } from './until';
 
 
 const generatedPhotos = Array.from({ length: photoValues.PHOTOS_COUNT }, createImage);
