--- js/randomInteger.js
@@ -1,7 +1,0 @@
-const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  return Math.floor(Math.random() * (upper - lower + 1)) + lower;
-};
-
-export {getRandomInteger};
